requirements.txt
requests
dnspython
xmltodict
jinja2
python-whois
